backend/requirements.txt
Django==4.2.14
django-filter==24.2
dj-database-url==2.2.0
graphene-django==3.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.3.2
pytest-django==4.8.0